fix: Fix Windows MCP stdout issue and libuv crash
- Created dedicated bin/seerxo-mcp.js and bin/seerxo.js entrypoints to isolate MCP stdio from TTY UI initialization.
- Removed aggressive process.exit() calls from command handlers. Now handlers correctly return and use process.exitCode = 1 on failure.
- Updated readline shutdown logic to track state and avoid duplicate rl.close() calls on already closed interfaces.
